Rename files to follow kebab-case convention
