+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-collection/edit/multi-product-picker.tsx
