+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-collection/edit/single-product-picker.tsx
